refactor(mcp): unify low/high-level tracing into one core
Both the low-level Server and high-level McpServer wrappers now funnel the
tool-call lifecycle through a single traceToolCall() in tracing-core.ts,
replacing two ~330-line parallel implementations. The list-tools and
initialize handlers also live in the core, removing the third list-tools
copy that lived in tools.ts.

Fixes that fall out of the single code path:
- analytics is fully isolated from the tool path: a failure preparing or
  publishing an event can no longer change what the tool returns or throws,
  and the tool's own errors are always re-thrown (was: publish-inside-try)
- unknown tools capture exactly one event instead of two
- the session id is resolved once per request and threaded through
  handleIdentify/getSessionInfo instead of being re-read from the shared,
  mutable data.sessionId (fixes the concurrent-request attribution race)

Net ~500 fewer lines. Behaviour-preserving; all 302 unit tests still pass.

Generated-By: PostHog Code
Task-Id: baa7e0cd-4946-4524-a05f-42c547a55f44
